refactor(tui): restore Work Window facade budget
